Removed jquery from actions.js
Replaced jQuery with vanilla js in actions.js in all instances except movement and animations. I'm going to refactor that to use SVG animations instead of JS, so I didn't spend the time converting that for now. jQuery should be gone everywhere else except a few instances on actions.js.